chat_channels/requirements.txt
beautifulsoup4==4.12.3
channels==4.1.0
daphne==4.1.2
Django==4.2.7
django_channels==0.7.0
faiss_cpu==1.8.0
langchain==0.3.7
langchain_community==0.3.7
langchain_openai==0.2.8
numpy==1.26.4
openai==1.54.4
Requests==2.32.3
setuptools
